divide wallet function
