+++ core/cli.go
@@ -112,20 +112,6 @@ func (cli *CLI) printChain() {
 		}
 	}
 }
-func (cli *CLI) createWallet() {
-	wallets := NewWallets()
-	address := wallets.CreateWallet()
-	wallets.SaveToFile()
-
-	fmt.Printf("address: %s\n", address)
-}
-
-func (cli *CLI) printWallets() {
-	wallets := NewWallets()
-	for address := range wallets.Wallets {
-		fmt.Printf("address: %s\n", address)
-	}
-}
 
 func (cli *CLI) printUtxo() {
 	up := GetUTXOPool()
