+++ core/cli_wallet.go
@@ -1,0 +1,18 @@
+package core
+
+import "fmt"
+
+func (cli *CLI) createWallet() {
+	wallets := NewWallets()
+	address := wallets.CreateWallet()
+	wallets.SaveToFile()
+
+	fmt.Printf("address: %s\n", address)
+}
+
+func (cli *CLI) printWallets() {
+	wallets := NewWallets()
+	for address := range wallets.Wallets {
+		fmt.Printf("address: %s\n", address)
+	}
+}
